refactor UserProfilePictureProp types from *.types.ts to it's main component file
